major_project/requirements.txt
matplotlib==3.8.1
numpy==1.25.2
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.76
Pillow==10.1.0
tensorflow==2.15.0
tensorflow_macos==2.15.0
watchdog==3.0.0
